backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-docx==1.1.2
pyyaml==6.0.2
google-genai
python-multipart==0.0.12
pydantic==2.9.0
python-dotenv==1.0.1
